Before, we’d allocate a new List, and then call ToArray on each token. The List is new, and each ToArray is a copy of the ReadOnlySpan representing the token.

In this PR, we clone the original input, and create Memory<T> pointing to each token.

We want the semantics that by calling ToList, the new is now entirely independent of the original input. Copies in all cases. This is to avoid gotchas where the underlying original data may be mutated. I think this is what users expect.