@clipperhouse 2024-06-21
@kevin-montrose @bretcope I don’t know if this is better, see description above. Is it better?