@clipperhouse 2024-06-21
A little faster, but more allocation:

```
after: 

| Method              | Mean       | Error    | StdDev   | Gen0     | Gen1     | Gen2     | Allocated   |
| ToList              | 1,484.0 us | 15.28 us | 14.29 us | 382.8125 | 373.0469 | 248.0469 | 2,155,937 B |

before:

| Method              | Mean       | Error    | StdDev   | Gen0     | Gen1     | Gen2     | Allocated   |
| ToList              | 1,847.3 us | 6.59 us  | 5.84 us  | 234.3750 | 201.1719 | 105.4688 | 1,522,633 B |
```