@clipperhouse 2024-06-21
Calling it: never mind, given the above results and the simplicity of the original version.